renames & fixes for ConcurrentListener
